Set LC_ALL inside container
For some reason, LC_ALL was unset. Not only did this cause warnings
durin the installation (each time perl was run), it also broke the
ceylon.formatter tests because @lucaswerkmeister is an asshole who puts
non-ASCII in his code just to spite everyone else working with his code
and then laughs at them when it breaks. What a dick.