Don’t copy over package sources from host
After the Debian Jessie release, `lxc-create -t debian` creates a Jessie
container, which doesn’t work with our Wheezy sources.

(According to the release notes [1] we could also have some other
problems, but I’m not seeing those. I suspect that in particular, the
lack of a problem with the init system is because we never actually run
an init system – this kernel doesn’t support lxc-attach, so we always
start the LXC with the required command, never actually booting it. I
think.)

[1]: https://www.debian.org/releases/jessie/i386/release-notes/ch-information.en.html#lxc-upgrade-issues